+++ build-logic/convention/src/main/java/AndroidApplicationConventionPlugin.kt
@@ -1,7 +1,7 @@
 import com.android.build.api.dsl.ApplicationExtension
 import com.android.build.api.dsl.Lint
-import dev.dai.android.architecture.template.build_logic.configure
-import dev.dai.android.architecture.template.build_logic.configureKotlinAndroid
+import dev.dai.android.architecture.template.buildlogic.configure
+import dev.dai.android.architecture.template.buildlogic.configureKotlinAndroid
 import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.gradle.kotlin.dsl.configure
