+++ build-logic/convention/src/main/java/AndroidLibraryComposeConventionPlugin.kt
@@ -1,5 +1,5 @@
 import com.android.build.gradle.LibraryExtension
-import dev.dai.android.architecture.template.build_logic.configureAndroidCompose
+import dev.dai.android.architecture.template.buildlogic.configureAndroidCompose
 import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.gradle.kotlin.dsl.apply
