+++ build-logic/convention/src/main/java/AndroidLibraryConventionPlugin.kt
@@ -1,9 +1,9 @@
 import com.android.build.api.dsl.Lint
 import com.android.build.api.variant.LibraryAndroidComponentsExtension
 import com.android.build.gradle.LibraryExtension
-import dev.dai.android.architecture.template.build_logic.configure
-import dev.dai.android.architecture.template.build_logic.configureKotlinAndroid
-import dev.dai.android.architecture.template.build_logic.disableUnnecessaryAndroidTests
+import dev.dai.android.architecture.template.buildlogic.configure
+import dev.dai.android.architecture.template.buildlogic.configureKotlinAndroid
+import dev.dai.android.architecture.template.buildlogic.disableUnnecessaryAndroidTests
 import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.gradle.kotlin.dsl.configure
