+++ build-logic/convention/src/main/java/dev/dai/android/architecture/template/buildlogic/KotlinAndroid.kt
@@ -1,4 +1,4 @@
-package dev.dai.android.architecture.template.build_logic
+package dev.dai.android.architecture.template.buildlogic
 
 import com.android.build.api.dsl.CommonExtension
 import org.gradle.api.JavaVersion
