+++ build-logic/convention/src/main/java/dev/dai/android/architecture/template/buildlogic/ProjectExtensions.kt
@@ -1,4 +1,4 @@
-package dev.dai.android.architecture.template.build_logic
+package dev.dai.android.architecture.template.buildlogic
 
 import org.gradle.api.Project
 import org.gradle.api.artifacts.MinimalExternalModuleDependency
